Tests/LibWeb/Text/expected/meta-color-scheme-invalidates-style.txt

Before: rgb(0, 0, 0)
After: rgb(235, 235, 235)
